VERSION 0.0.1 bumped version to move from CVS to git
Updated tino/ and bump VERSION just for git-log purpose

Please note that this now is old.  Modern kernels seem to have
completely eliminated support for hardware buzzers, which is very sad!

------------------------------------------------------------------------

VERSION 0.0.0

This actually is an old (but now improved) version.

I just created it to let my console beep.